+++ lib/PHPParser/Node/Stmt/Class.php
@@ -68,6 +68,16 @@ class PHPParser_Node_Stmt_Class extends PHPParser_Node_Stmt
         return (bool) ($this->type & self::MODIFIER_FINAL);
     }
 
+    public function getMethods() {
+        $methods = array();
+        foreach ($this->stmts as $stmt) {
+            if ($stmt instanceof PHPParser_Node_Stmt_ClassMethod) {
+                $methods[] = $stmt;
+            }
+        }
+        return $methods;
+    }
+
     public static function verifyModifier($a, $b) {
         if ($a & 7 && $b & 7) {
             throw new PHPParser_Error('Multiple access type modifiers are not allowed');